--DB version control: Initial Commit--
(this version defines and uses version 0.1 of the database)

Created a database table (DB_UPGRADE_LOGS) that stores the upgrade history for a given database instance so it is clear what each database instance's version is.  This is used to apply the necessary database upgrades in order to deploy a given version of an associated application.