--DB version control: Documentation/Template Update--
(this version uses version 0.1 of the database)

Compiled a new document DB_version_control\docs\DB Version Control Module SOP.docx that contains the information/procedures necessary to implement the Database Version Control Module.

Updated the DB_version_control\SQL\README.txt to include all information from the new SOP document.  This was copied to each of the template directories so it can be easily included in a new/existing code repository.

Renamed DB_version_control\docs\commit_message_info.txt to DB_version_control_templates.txt and updated with template placeholders and a detailed explanation for each placeholder.  Also added section headings for organizational purposes. This was copied to each of the template_directories so it can be easily included in a new/existing code repository.

Created directory templates for the shared database and dedicated database cases in the docs\repository_templates folder and included the README.txt and DB_version_control_templates.txt files so the DB version control module folder structure and template files can be added easily to a given repository.  Also, created simple templates for a sample upgrade file and a sample combined DDL/DML file.

Developed a new diagram for the DB Version Control Module's directory structure for the new SOP equivalent to the simple diagram in the README.txt file.

Updated the DB Version Control Module Documentation.docx with a link to the new SOP.